refactor(notification): extract Zod envelope to its own file and drop versioning
Two related cleanups to make the outbox payload contract self-evident
and trim transitional code that was never deployed:

- Move the structural Zod validator from the inside of NotificationOutboxService
  to a dedicated `notification.schema.ts`. Its job is JSON-boundary structural
  validation; the rich `NotificationJobData` TS interface continues to provide
  semantic typing (notably `templateData`) for downstream consumers. They live
  at intentionally different abstraction levels.
- Add `notification-job-data.contract.spec.ts` with a compile-time mutual
  assignability assertion on the envelope fields (everything except the two
  intentionally-divergent `recipients`/`templateData` slots). Drift on
  id/type/channels/bookingId/pushPayload/priority fails the test at type-check
  time.
- Drop the `OUTBOX_SCHEMA_VERSION` envelope and the `legacyOutboxPayloadSchema`
  v1 fallback. The (eventType, subtype) pair is the discriminator; nothing has
  shipped that wrote a v1 payload, so the back-compat path was never live.
  Reintroduce versioning when there's an actual breaking change to evolve
  through.

Refs: notification-outbox extensibility review (3B + 7-aggressive).
